backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0
httpx==0.25.2
python-dotenv==1.0.0
cryptography==41.0.7